Move publish with selector operator to Publish.kt file
