+++ tools/binary-compatibility/reference-public-api/reaktive.txt
@@ -596,9 +596,6 @@ public final class com/badoo/reaktive/observable/OnErrorReturnKt {
 
 public final class com/badoo/reaktive/observable/PublishKt {
 	public static final fun publish (Lcom/badoo/reaktive/observable/Observable;)Lcom/badoo/reaktive/observable/ConnectableObservable;
-}
-
-public final class com/badoo/reaktive/observable/PublishWithSelectorKt {
 	public static final fun publish (Lcom/badoo/reaktive/observable/Observable;Lkotlin/jvm/functions/Function1;)Lcom/badoo/reaktive/observable/Observable;
 }
